# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/sync-runs": {
      "post": {
        "operationId": "SyncController_start",
        "parameters": [],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or revoked API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Valid key, wrong scope (a device key on a user endpoint)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start a Kindle sync (or get the one already in progress); poll it for status",
        "tags": [
          "sync"
        ]
      },
      "get": {
        "operationId": "SyncController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SyncRunDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (details has one message per field)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or revoked API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Valid key, wrong scope (a device key on a user endpoint)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Your sync runs, newest first",
        "tags": [
          "sync"
        ]
      }
    },
    "/v1/sync-runs/{id}": {
      "get": {
        "operationId": "SyncController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (details has one message per field)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or revoked API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Valid key, wrong scope (a device key on a user endpoint)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or it belongs to someone else (we don't say which)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "sync"
        ]
      }
    }
  },
  "info": {
    "title": "Stacks API",
    "description": "Reading tracker. Every route needs `Authorization: Bearer <key>`. Errors are always `{ error: { code, message, details? } }`.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "not_found"
          },
          "message": {
            "type": "string",
            "example": "Library book not found"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "name": "must be 1–80 characters"
            }
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorBody": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "SyncRunDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "device",
              "kindle_sim"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "description": "queued → running → succeeded | failed"
          },
          "eventsIngested": {
            "type": "number",
            "description": "Events recorded so far (counts up while running)"
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "source",
          "status",
          "eventsIngested",
          "error",
          "createdAt",
          "startedAt",
          "finishedAt"
        ]
      }
    }
  }
}